Remove temporary mountpoint only if it exists
If the temporary mountpoint does not exist, we should be happy, since
there is nothing to do. Unfortunately rmdir (unlike rm -Rf) reports an
error in this case, so test for the directory existence before.

+++ usr/share/rear/lib/global-functions.sh
@@ -632,7 +632,9 @@ function mount_url() {
 }
 
 function remove_temporary_mountpoint() {
-    rmdir $v "$1" >&2
+    if test -d "$1" ; then
+        rmdir $v "$1" >&2
+    fi
 }
 
 ### Unmount url $1 at mountpoint $2, perform mountpoint cleanup and exit task + error handling
